+++ packages/core/src/error/formatter.ts
@@ -1,9 +1,9 @@
+import type { ProcedureType } from '../procedure';
 import type {
   TRPC_ERROR_CODE_KEY,
   TRPC_ERROR_CODE_NUMBER,
   TRPCErrorShape,
 } from '../rpc';
-import type { ProcedureType } from '../types';
 import type { TRPCError } from './TRPCError';
 
 /**
