+++ packages/core/src/types.ts
@@ -19,12 +19,6 @@ export type inferRouterError<TRouter extends AnyRouter> =
 export type inferRouterMeta<TRouter extends AnyRouter> =
   inferRouterConfig<TRouter>['$types']['meta'];
 
-export const procedureTypes = ['query', 'mutation', 'subscription'] as const;
-/**
- * @public
- */
-export type ProcedureType = (typeof procedureTypes)[number];
-
 export type inferHandlerInput<TProcedure extends AnyProcedure> = ProcedureArgs<
   inferProcedureParams<TProcedure>
 >;
